requirements.txt
Flask==2.3.5
python-telegram-bot==20.3
qrcode==7.4.2